requirements.txt
streamlit
requests
beautifulsoup4
google-generativeai
faiss-cpu
numpy
python-dotenv
